[#119893127] Reorganise folder structure
This will allow for easier refactoring and grouping as the
application grows.
